refactor(hosts): extract managed skill-facade file handling
