+++ loopx/canary/module_metric_baseline.json
@@ -15,11 +15,6 @@
       "dict_any_count": 0,
       "lines": 1992
     },
-    "loopx/slash_command_install.py": {
-      "any_count": 11,
-      "dict_any_count": 0,
-      "lines": 1514
-    },
     "loopx/capabilities/auto_research/demo_e2e.py": {
       "any_count": 0,
       "dict_any_count": 0,
