+++ tests/test_gemini_cursor_host_surfaces.py
@@ -9,13 +9,15 @@ the facade dead-ends at argparse and the surface is decorative.
 
 from __future__ import annotations
 
-import json
-import shlex
-import subprocess
-import sys
 from pathlib import Path
 
 import pytest
+
+from host_surface_cli_probes import (
+    onboarding_setup_command_installs,
+    selection_gate_offers_surface,
+    start_goal_accepts_surface,
+)
 
 from loopx.agent_onboarding import _start_instruction, _surface_install_command
 from loopx.host_loop_activation import (
@@ -28,102 +30,21 @@ from loopx.host_loop_activation import (
 NEW_HOSTS = ("gemini-cli", "cursor-agent")
 
 
-def _run_cli(*args: str, cwd: Path, env: dict[str, str] | None = None) -> subprocess.CompletedProcess[str]:
-    return subprocess.run(
-        [sys.executable, "-m", "loopx.cli", "--format", "json", *args],
-        cwd=cwd,
-        check=False,
-        text=True,
-        capture_output=True,
-        env=env,
-    )
-
-
-def _connected_project(root: Path, goal_id: str = "surface-goal") -> Path:
-    project = root / "project"
-    registry_path = project / ".loopx" / "registry.json"
-    registry_path.parent.mkdir(parents=True)
-    registry_path.write_text(
-        json.dumps(
-            {
-                "goals": [
-                    {
-                        "id": goal_id,
-                        "domain": "test",
-                        "status": "active",
-                        "repo": str(project),
-                        "adapter": {
-                            "kind": "generic_project_goal_v0",
-                            "status": "connected",
-                        },
-                    }
-                ]
-            }
-        ),
-        encoding="utf-8",
-    )
-    return project
-
-
 @pytest.mark.parametrize("host_surface", NEW_HOSTS)
 def test_start_goal_accepts_the_new_host_surface(tmp_path: Path, host_surface: str) -> None:
-    """The exact command the installed facade generates must run, not exit 2."""
-    project = _connected_project(tmp_path)
-    result = _run_cli(
-        "start-goal",
-        "--guided",
-        "--project",
-        str(project),
-        "--goal-id",
-        "surface-goal",
-        "--host-surface",
-        host_surface,
-        "--goal-text",
-        "verify the host contract for this surface",
-        cwd=tmp_path,
-    )
-    assert result.returncode == 0, result.stderr
-    payload = json.loads(result.stdout)
-    assert payload["ok"] is True
-    assert payload["host_surface"] == host_surface
+    payload = start_goal_accepts_surface(host_surface, tmp_path)
     activation = payload["command_pack"]["host_loop_activation"]
-    assert activation["agent_type"] == host_surface
     assert activation["host_surface"] == {
         "gemini-cli": "gemini_cli_agent_loop",
         "cursor-agent": "cursor_agent_loop",
     }[host_surface]
 
 
+@pytest.mark.parametrize("host_surface", NEW_HOSTS)
 def test_host_selection_gate_offers_the_new_surfaces_and_its_rerun_command_works(
-    tmp_path: Path,
+    tmp_path: Path, host_surface: str
 ) -> None:
-    """The facade falls back to the selection gate when the host is unclear, so
-    a host missing from the gate is unreachable even though it exists."""
-    project = _connected_project(tmp_path)
-    gate_result = _run_cli(
-        "start-goal",
-        "--guided",
-        "--project",
-        str(project),
-        "--goal-id",
-        "surface-goal",
-        "--goal-text",
-        "verify the host selection gate",
-        cwd=tmp_path,
-    )
-    assert gate_result.returncode == 0, gate_result.stderr
-    gate = json.loads(gate_result.stdout)["host_surface_selection_gate"]
-    choices = {item["host_surface"]: item for item in gate["choices"]}
-    for host_surface in NEW_HOSTS:
-        assert host_surface in choices, sorted(choices)
-
-    # Run one offered choice exactly as printed: the gate is only useful if its
-    # rerun_command is executable as-is.
-    tokens = shlex.split(choices["cursor-agent"]["rerun_command"])
-    assert tokens[0] == "loopx"
-    rerun = _run_cli(*tokens[1:], cwd=tmp_path)
-    assert rerun.returncode == 0, rerun.stderr
-    assert json.loads(rerun.stdout)["host_surface"] == "cursor-agent"
+    selection_gate_offers_surface(host_surface, tmp_path)
 
 
 @pytest.mark.parametrize("agent_type", NEW_HOSTS)
@@ -135,7 +56,6 @@ def test_agent_onboarding_setup_command_installs_that_surface(
     """agent-onboard hands back a setup command; executing it must provision the
     host it named, from any cwd."""
     monkeypatch.delenv("LOOPX_SKILLS_DIR", raising=False)
-    project = _connected_project(tmp_path)
     outside = tmp_path / "outside"
     outside.mkdir()
     env = {
@@ -145,34 +65,13 @@ def test_agent_onboarding_setup_command_installs_that_surface(
         "CURSOR_HOME": str(tmp_path / "cursor"),
     }
 
-    onboard = _run_cli(
-        "agent-onboard",
-        "--agent-type",
-        agent_type,
-        "--project",
-        str(project),
-        "--goal-id",
-        "surface-goal",
-        cwd=outside,
-        env=env,
-    )
-    assert onboard.returncode == 0, onboard.stderr
-    payload = json.loads(onboard.stdout)
-    assert payload["agent_type"] == agent_type
-    # These hosts get their skills from the LoopX installer, not from a host
-    # that manages skills itself.
-    assert payload["skill_delivery"]["mode"] == "surface_managed"
-
-    facade = payload["commands"]["install_command_facade"]
-    assert facade is not None
-    tokens = shlex.split(facade)
-    assert tokens[0] == "loopx"
-    install = _run_cli(*tokens[1:], cwd=outside, env=env)
-    assert install.returncode == 0, install.stderr
-    assert json.loads(install.stdout)["ok"] is True
-
     home = tmp_path / ("gemini" if agent_type == "gemini-cli" else "cursor")
-    assert (home / "skills" / "loopx" / "SKILL.md").is_file()
+    onboarding_setup_command_installs(
+        agent_type,
+        outside,
+        env,
+        expected_skill=home / "skills" / "loopx" / "SKILL.md",
+    )
 
 
 @pytest.mark.parametrize("agent_type", NEW_HOSTS)
